feat: implement Firebase token verification middleware and refactor header components
